# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  initials: string;
  accent: string;
}

export type DocumentListItemAccess = typeof DocumentListItemAccess[keyof typeof DocumentListItemAccess];


export const DocumentListItemAccess = {
  owned: 'owned',
  shared: 'shared',
} as const;

export interface DocumentListItem {
  id: string;
  title: string;
  ownerId: string;
  ownerName: string;
  updatedAt: string;
  wordCount: number;
  access: DocumentListItemAccess;
  preview: string;
}

export type DocumentAccess = typeof DocumentAccess[keyof typeof DocumentAccess];


export const DocumentAccess = {
  owned: 'owned',
  shared: 'shared',
} as const;

export interface Share {
  userId: string;
  userName: string;
  userEmail: string;
  initials: string;
  accent: string;
  sharedAt: string;
}

export interface Document {
  id: string;
  title: string;
  content: string;
  ownerId: string;
  ownerName: string;
  updatedAt: string;
  createdAt: string;
  wordCount: number;
  access: DocumentAccess;
  shares: Share[];
}

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  content: string;
}

export interface DocumentUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
}

export interface DocumentImportInput {
  /** @minLength 1 */
  filename: string;
  content: string;
}

export interface ShareInput {
  /** @minLength 1 */
  userId: string;
}

export interface Dashboard {
  ownedCount: number;
  sharedCount: number;
  totalWords: number;
  recent: DocumentListItem[];
}

export interface Error {
  error: string;
}

export type UserIdHeaderParameter = string;