# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentListItem } from './documentListItem';

export interface Dashboard {
  ownedCount: number;
  sharedCount: number;
  totalWords: number;
  recent: DocumentListItem[];
}