# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentAccess } from './documentAccess';
import type { Share } from './share';

export interface Document {
  id: string;
  title: string;
  content: string;
  ownerId: string;
  ownerName: string;
  updatedAt: Date;
  createdAt: Date;
  wordCount: number;
  access: DocumentAccess;
  shares: Share[];
}