# lib/api-zod/src/generated/types/documentImportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentImportInput {
  /** @minLength 1 */
  filename: string;
  content: string;
}