# lib/api-zod/src/generated/types/documentListItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentListItemAccess } from './documentListItemAccess';

export interface DocumentListItem {
  id: string;
  title: string;
  ownerId: string;
  ownerName: string;
  updatedAt: Date;
  wordCount: number;
  access: DocumentListItemAccess;
  preview: string;
}