# lib/api-zod/src/generated/types/share.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Share {
  userId: string;
  userName: string;
  userEmail: string;
  initials: string;
  accent: string;
  sharedAt: Date;
}